feat: created comp for onboarding
